Reorganize codebase into packages (#119)
* Reorganize codebase

* lint

* reorganize test files

* move add_indices back to internals

* simplify import and rename test file

* reorganize docs

* propagate changes to notebook imports

* remove gallery
