+++ docs/source/counterfactual.rst
@@ -1,16 +1,35 @@
-Counterfactual Handlers
-=======================
+Counterfactual
+==============
 
 .. automodule:: causal_pyro.counterfactual
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+
+Operations
+----------
+
+.. automodule:: causal_pyro.counterfactual.ops
+   :members:
+   :undoc-members:
+
+Handlers
+--------
 
 .. automodule:: causal_pyro.counterfactual.handlers
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+
+.. automodule:: causal_pyro.counterfactual.handlers.ambiguity
+   :members:
+   :undoc-members:
+
+.. automodule:: causal_pyro.counterfactual.handlers.selection
+   :members:
+   :undoc-members:
+
+Internals
+---------
 
 .. automodule:: causal_pyro.counterfactual.internals
-    :members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
