+++ docs/source/indexed.rst
@@ -1,0 +1,28 @@
+Indexed
+=======
+
+.. automodule:: causal_pyro.indexed
+   :members:
+   :undoc-members:
+
+
+Operations
+----------
+
+.. automodule:: causal_pyro.indexed.ops
+   :members:
+   :undoc-members:
+
+Handlers
+--------
+
+.. automodule:: causal_pyro.indexed.handlers
+   :members:
+   :undoc-members:
+
+Internals
+---------
+
+.. automodule:: causal_pyro.indexed.internals
+   :members:
+   :undoc-members:
