+++ docs/source/interventional.rst
@@ -1,0 +1,20 @@
+Interventional
+==============
+
+.. automodule:: causal_pyro.interventional
+   :members:
+   :undoc-members:
+
+Operations
+----------
+
+.. automodule:: causal_pyro.interventional.ops
+   :members:
+   :undoc-members:
+
+Handlers
+--------
+
+.. automodule:: causal_pyro.interventional.handlers
+   :members:
+   :undoc-members:
