+++ docs/source/observational.rst
@@ -1,0 +1,13 @@
+Observational
+=============
+
+.. automodule:: causal_pyro.observational
+   :members:
+   :undoc-members:
+
+Handlers
+--------
+
+.. automodule:: causal_pyro.observational.handlers
+   :members:
+   :undoc-members:
