--- docs/source/primitives.rst
@@ -1,5 +1,0 @@
-Primitive interventions
-=======================
-
-.. automodule:: causal_pyro.primitives
-   :members:
