--- docs/source/query.rst
@@ -1,12 +1,0 @@
-Causal Query Operations
-=======================
-
-.. automodule:: causal_pyro.query
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-.. automodule:: causal_pyro.query.do_messenger
-    :members:
-    :undoc-members:
-    :show-inheritance:
