--- docs/source/reparam.rst
@@ -1,12 +1,0 @@
-Reparameterizers
-==================
-
-.. automodule:: causal_pyro.reparam
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-.. automodule:: causal_pyro.reparam.soft_conditioning
-    :members:
-    :undoc-members:
-    :show-inheritance:
