+++ tests/counterfactual/test_ambiguity.py
@@ -5,13 +5,16 @@ import pyro.distributions as dist
 import pytest
 import torch
 
-from causal_pyro.counterfactual.handlers import (
+from causal_pyro.counterfactual.handlers.selection import (
+    SelectCounterfactual,
+    SelectFactual,
+)
+from causal_pyro.counterfactual.ops import (
     Factual,
     MultiWorldCounterfactual,
     TwinWorldCounterfactual,
 )
-from causal_pyro.counterfactual.selection import SelectCounterfactual, SelectFactual
-from causal_pyro.query.do_messenger import do
+from causal_pyro.interventional.handlers import do
 
 logger = logging.getLogger(__name__)
