+++ tests/counterfactual/test_do_messenger.py
@@ -5,14 +5,14 @@ import pyro.distributions as dist
 import pytest
 import torch
 
-from causal_pyro.counterfactual.handlers import (
+from causal_pyro.counterfactual.ops import (
     BaseCounterfactual,
     Factual,
     MultiWorldCounterfactual,
     TwinWorldCounterfactual,
 )
-from causal_pyro.primitives import intervene
-from causal_pyro.query.do_messenger import do
+from causal_pyro.interventional.handlers import do
+from causal_pyro.interventional.ops import intervene
 
 logger = logging.getLogger(__name__)
