+++ tests/counterfactual/test_mediation.py
@@ -7,11 +7,11 @@ import pyro.distributions as dist
 import pytest
 import torch
 
-from causal_pyro.counterfactual.handlers import (
+from causal_pyro.counterfactual.ops import (
     MultiWorldCounterfactual,
     TwinWorldCounterfactual,
 )
-from causal_pyro.query.do_messenger import DoMessenger, do
+from causal_pyro.interventional.handlers import DoMessenger, do
 
 logger = logging.getLogger(__name__)
