+++ tests/counterfactual/test_selection.py
@@ -5,12 +5,15 @@ import pyro.distributions as dist
 import pytest
 import torch
 
-from causal_pyro.counterfactual.handlers import (
+from causal_pyro.counterfactual.handlers.selection import (
+    SelectCounterfactual,
+    SelectFactual,
+)
+from causal_pyro.counterfactual.ops import (
     MultiWorldCounterfactual,
     TwinWorldCounterfactual,
 )
-from causal_pyro.counterfactual.selection import SelectCounterfactual, SelectFactual
-from causal_pyro.query.do_messenger import do
+from causal_pyro.interventional.handlers import do
 
 logger = logging.getLogger(__name__)
