+++ tests/indexed/test_index_set.py
@@ -2,7 +2,7 @@ import logging
 
 import pytest
 
-from causal_pyro.primitives import IndexSet, union
+from causal_pyro.indexed.ops import IndexSet, union
 
 logger = logging.getLogger(__name__)
